requirements.txt
sounddevice
numpy
SpeechRecognition
googletrans
pyttsx3
pygame

openai>=1.0.0
groq

openai-whisper>=20231117
librosa>=0.10.0

pyaudio>=0.2.11  
PyAudio
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0